Simplify the cell metadata load+save operations.
This removes all of the conditional logic regarding cell state in
Grid.save() and Grid.load(), by changing the cell metadata storage to
always be in the puzfile format.  This also means if someone were to add
another value some day, it won't break our assumptions about the values.

This patch also changes it so we always read the cell metadata through
properties and set change it through accessors.

Signed-off-by: Peter Jones <pjones@redhat.com>